Actualizo tipo de dato

+++ ProyectoFinal_CristianLopez/Back-End/src/main/java/com/portfolio/CristianLopez/Security/Repository/UsuarioRepository.java
@@ -14,11 +14,12 @@ import org.springframework.stereotype.Repository;
  * @author krono
  */
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long>{
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer>{
     Optional <Usuario> findByUsername(String username);
-    
+   
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
     
+   
 
 }
